Get address page and initial console page

+++ api/src/main/java/mempool/space/model/Stats.kt
@@ -5,9 +5,9 @@ import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
 data class Stats(
-    @Json(name= "funded_txo_count") val fundedTxoCount: Int,
-    @Json(name= "funded_txo_sum") val fundedTxoSum: Int,
-    @Json(name= "spent_txo_count") val spentTxoCount: Int,
-    @Json(name= "spent_txo_sum") val spentTxoSum: Int,
-    @Json(name= "tx_count") val txCount: Int,
+    @Json(name= "funded_txo_count") val fundedTxoCount: Long,
+    @Json(name= "funded_txo_sum") val fundedTxoSum: Long,
+    @Json(name= "spent_txo_count") val spentTxoCount: Long,
+    @Json(name= "spent_txo_sum") val spentTxoSum: Long,
+    @Json(name= "tx_count") val txCount: Long,
 )
